+++ test/main.cairo
@@ -1,36 +1,31 @@
 %lang starknet
-from starkware.cairo.common.math import assert_nn, assert_not_zero
 from starkware.cairo.common.cairo_builtins import HashBuiltin
+
+from test.types import ImportedStruct
 
 struct MyStruct {
     index: felt,
 }
 
 @storage_var
-func balance() -> (res: felt) {
+func storage_skipped() -> (res: felt) {
 }
 
 @external
-func increase_balance{syscall_ptr: felt*, pedersen_ptr: HashBuiltin*, range_check_ptr}(
-    amount: felt, test_len: felt, test: MyStruct*
+func struct_in_arg{syscall_ptr: felt*, pedersen_ptr: HashBuiltin*, range_check_ptr}(
+    amount: felt, _struct: MyStruct, array_len: felt, array: felt*
 ) {
-    with_attr error_message("Amount must be positive. Got: {amount}.") {
-        assert_nn(amount);
-    }
-
-    let (res) = balance.read();
-    balance.write(res + amount);
     return ();
 }
 
 @view
-func get_balance{syscall_ptr: felt*, pedersen_ptr: HashBuiltin*, range_check_ptr}() -> (res: felt) {
-    let (res) = balance.read();
-    return (res,);
+func struct_ptr_in_return{syscall_ptr: felt*, pedersen_ptr: HashBuiltin*, range_check_ptr}() -> (
+    res_len: felt, res: felt*, arr_len: felt, arr: ImportedStruct*
+) {
+    return (1, new (1), 1, new ImportedStruct(0));
 }
 
 @constructor
 func constructor{syscall_ptr: felt*, pedersen_ptr: HashBuiltin*, range_check_ptr}() {
-    balance.write(0);
     return ();
 }
