+++ test/types.cairo
@@ -1,0 +1,3 @@
+struct ImportedStruct {
+    index: felt,
+}
